Move Mood class to mood.py
